1.兼容最低API16 2.更新友盟统计sdk 3.解决一日通页面无数据显示空白的问题 4.解决检查更新下载异常的问题 5.修改字段

+++ app/src/main/java/com/pictureair/photopass/greendao/PictureAirDbManager.java
@@ -870,7 +870,7 @@ public class PictureAirDbManager {
      * @param threadId
      * @param finished
      */
-    public static void updateThread(String url, int threadId, int finished) {
+    public static void updateThread(String url, int threadId, long finished) {
         ThreadInfoDao threadInfoDao = MyApplication.getInstance().getDaoSession().getThreadInfoDao();
         QueryBuilder<ThreadInfo> queryBuilder = threadInfoDao.queryBuilder()
                 .where(ThreadInfoDao.Properties.Url.eq(url), ThreadInfoDao.Properties.ThreadId.eq(threadId));
